backend/requirements.txt
Flask
SQLAlchemy
Werkzeug
Flask-Cors
zxcvbn-python